pyproject.toml
[project]
name = "deribit-price-tracker"
version = "0.1.0"
description = "Трекер цен криптовалют с Deribit"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.46",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "pydantic-settings>=2.13.0",
    "aiohttp>=3.9.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
]
